moving blog menu to the header
