+++ src/blog-components/blog-footer/BlogFooter.tsx
@@ -7,6 +7,20 @@ export interface BlogFooterProps {
 const BlogFooter = ({ data }: BlogFooterProps) => (
   <footer className="blog-footer">
     <p className="bloog-footer-text">{data.text}</p>
+
+    <nav className="flex space-x-6">
+      {data.socialLinks.map((headerItem) => (
+        <a
+          key={headerItem.url}
+          href={headerItem.url}
+          className="header-social-link"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          <headerItem.socialIcon />
+        </a>
+      ))}
+    </nav>
   </footer>
 );
